// src/shared/enums/FoafOptions.ts
enum FoafOptions {
  ACCOUNT = 'Online account',
  AGE = 'Age',
  AIM_CHAT_ID = 'AIM chat ID',
  BASED_NEAR = 'Based near',
  BIRTHDAY = 'Birthday',
  CURRENT_PROJECT = 'Current project',
  DEPICTION = 'Profile picture',
  FAMILY_NAME = 'Family name',
  FIRST_NAME = 'First name',
  GENDER = 'Gender',
  GIVEN_NAME = 'Given name',
  HOMEPAGE = 'Homepage',
  ICQ_CHAT_ID = 'ICQ chat ID',
  IMAGES = 'Images',
  INTERESTS = 'Interests',
  JABBER_ID = 'Jabber ID',
  KNOWS = 'Knows',
  LAST_NAME = 'Last name',
  CREATED = 'Created',
  EMAIL = 'E-mail address',
  MSN_CHAT_ID = 'MSN chat ID',
  MYERS_BRIGGS = 'Myers Briggs',
  NAME = 'Name',
  NICKNAME = 'Nickname',
  OPEN_ID = 'Open ID',
  PAGE = 'Page',
  PAST_PROJECT = 'Past project',
  PHONE = 'Phone',
  PUBLICATIONS = 'Publications',
  SCHOOL_HOMEPAGE = 'School homepage',
  SKYPE_ID = 'Skype ID',
  SURNAME = 'Surname',
  TIP_JAR = 'Tip jar',
  TITLE = 'Title',
  WEBLOG = 'Web blog',
  WORK_INFO_HOMEPAGE = 'Work info homepage',
  WORKPLACE_HOMEPAGE = 'Workplace homepage',
  YAHOO_CHAT_ID = 'Yahoo chat ID'
}

export default FoafOptions;